requirements.txt
# Required python packages.
# Can be installed py "pip install -r requirements.txt"

django-autoslug
django-taggit
